cli/requirements.txt
paddlepaddle==2.3.2
paddlenlp==2.4.5